Mejorar el método delete para el manejo de errores

+++ src/main/java/com/mobydigital/evaluacion/service/TurnoService.java
@@ -57,6 +57,9 @@ public class TurnoService implements ITurnoService{
 
     @Override
     public void deleteTurno(Long id) {
+        if (!turnoRepository.existsById(id)) { //Verifico si no existe el turno buscado
+            throw new RecursoNoEncontradoException("El turno con ID " + id + " no existe. ");
+        }
         turnoRepository.deleteById(id);
     }
 }
